refactor(latexer): move common converters to common.py
